Several fixes for when there are already events in the input to Evita.
Evita would throw errors when it tried to add an event to chunks that already have events. This was fixed by not trying to look for an event in a noun chunk if the head is an event. Fixing this actually required fixing a little cascade of small errors that had never popped up before.

In addition, while fixing this I made sure that the common Blinker errors for timex linking are always intercepted, and actually logged as a warning.