Simplified how Tags are created.
This was triggered by unclaities on how the 'id' attribute was used or abused). Part of this was to fix issue #56, but it turned out that there was no use at all for the 'id' attribute on Tags so got completely rid of it. Left some notes in docs/notes/creating-tags.md.